requirements.txt
streamlit==1.31.0
google-api-python-client==2.108.0
nltk==3.8.1
pandas==2.1.4
plotly==5.18.0
python-dotenv==1.0.0
scipy==1.11.4
